feat: move kafka consumer logic to use case
